adding status code and message

+++ src/modules/expense-sub-category/expense-sub-category.service.ts
@@ -1,10 +1,14 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { FindManyOptions, Repository } from 'typeorm';
 import { ExpenseSubCategory } from './entities/expense-sub-category.entity';
 import { ExpenseCategory } from '../expense-category/entities/expense-category.entity';
 import { CreateExpenseSubCategoryInput } from './dto/create-expense-sub-category.input';
 import { UpdateExpenseSubCategoryInput } from './dto/update-expense-sub-category.input';
+import {
+  CommonExpenseSubCategoryResponse,
+  PaginatedExpenseSubCategoryResponse,
+} from './dto/expense-sub-category-response.dto';
 
 @Injectable()
 export class ExpenseSubCategoryService {
@@ -17,7 +21,7 @@ export class ExpenseSubCategoryService {
 
   async create(
     createExpenseSubCategoryInput: CreateExpenseSubCategoryInput,
-  ): Promise<ExpenseSubCategory> {
+  ): Promise<CommonExpenseSubCategoryResponse> {
     const expenseCategory = await this.expenseCategoryRepository.findOneBy({
       id: createExpenseSubCategoryInput.expenseCategoryId,
     });
@@ -33,26 +37,59 @@ export class ExpenseSubCategoryService {
       isActive: createExpenseSubCategoryInput.isActive,
     });
 
-    return this.expenseSubCategoryRepository.save(expenseSubCategory);
+    await this.expenseSubCategoryRepository.save(expenseSubCategory);
+
+    return {
+      statusCode: 201,
+      message: 'Expense Sub Category created successfully',
+    };
   }
 
-  findAll(): Promise<ExpenseSubCategory[]> {
-    return this.expenseSubCategoryRepository.find({
-      relations: ['expenseCategory', 'expenses'],
-    });
+  async findAll(
+    perPage: number,
+    page: number,
+  ): Promise<PaginatedExpenseSubCategoryResponse> {
+    const options: FindManyOptions<ExpenseSubCategory> = {
+      take: perPage,
+      skip: (page - 1) * perPage,
+      relations: ['expenseCategory'],
+      order: { id: 'DESC' },
+    };
+    const [data, totalRows] =
+      await this.expenseSubCategoryRepository.findAndCount(options);
+    const totalPages = Math.ceil(totalRows / perPage);
+
+    return {
+      data,
+      totalRows,
+      totalPages,
+      currentPage: page,
+      statusCode: 200,
+      message: 'Expenses Sub Category fetched successfully',
+    };
   }
 
-  async findOne(id: number): Promise<ExpenseSubCategory> {
-    return await this.expenseSubCategoryRepository.findOneOrFail({
+  async findOne(id: number): Promise<CommonExpenseSubCategoryResponse> {
+    const data = await this.expenseSubCategoryRepository.findOne({
       where: { id },
       relations: ['expenseCategory', 'expenses'],
     });
+    if (!data) {
+      throw new NotFoundException(
+        `Expense Subcategory with ID ${id} was not found`,
+      );
+    }
+    return {
+      data,
+      statusCode: 200,
+      message: 'Expense Sub category fetched successfully',
+    };
   }
 
   async update(
     id: number,
     updateExpenseSubCategoryInput: UpdateExpenseSubCategoryInput,
-  ): Promise<ExpenseSubCategory> {
+  ): Promise<CommonExpenseSubCategoryResponse> {
     const { expenseCategoryId, ...updateData } = updateExpenseSubCategoryInput;
 
     const expenseSubCategory =
@@ -67,11 +104,22 @@ export class ExpenseSubCategoryService {
     }
 
     Object.assign(expenseSubCategory, updateData);
-    return this.expenseSubCategoryRepository.save(expenseSubCategory);
+    await this.expenseSubCategoryRepository.save(expenseSubCategory);
+    return {
+      statusCode: 200,
+      message: 'Expense updated successfully',
+    };
   }
-  async remove(id: number): Promise<ExpenseSubCategory> {
-    const expenseSubCategory = await this.findOne(id);
+  async remove(id: number): Promise<CommonExpenseSubCategoryResponse> {
+    const expenseSubCategory =
+      await this.expenseSubCategoryRepository.findOneBy({ id: id });
+    if (!expenseSubCategory) {
+      throw new NotFoundException(`Expense Category with ID ${id} not found`);
+    }
     await this.expenseSubCategoryRepository.remove(expenseSubCategory);
-    return { ...expenseSubCategory, id };
+    return {
+      statusCode: 200,
+      message: 'Expense deleted successfully',
+    };
   }
 }
